backend/requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
Flask-Cors
googlemaps